Changed account number, BSB and ABN to type string.
There are several banks in Australia and New Zealand which have a 0 at the
front of their BSB numbers eg Commonwealth Bank. With the type as an
integer it drops the leading 0 which gives a validation error when sent to
the server as the BSB is too short.

The first two digits of an ABN number is a checksum which can also contain
a zero.

You're not going to add a bank account to another bank account number and
divide it by their BSB and times it by their ABN.
Rule of thumb: If you not doing arithmetic with the numbers, store them as
a string.