Contacts Endpoint - Method to find multiple
Similar to how InvoicesEndpoint has an Ids method to find multiple
Invoices at the same time, this method is for retreiving multiple
contacts at the same time which helps with batching requests and not
hitting the API request limit as quickly when fetching a lot of data.

Also added a unit test method for listing by ids and that passes.