While condition is always true the second time
I changed the `while` to an `if` because the previous implementation would always return `false` the second time that condition was checked. The queue is dequeued with each enqueue and all the properties involved are private, so it's safe to assume the condition only needs to be checked once. Note, this is not thread safe, but RateLimiter already is not thread safe (https://github.com/XeroAPI/Xero-Net/issues/248).